Intraday feature update.
Developer give a score for each indicator for a moment every day.
User can choose a day and show the indicators and what score and why the
developer gave at that moment.
The score and the reason save in .xlsx file for each stock.

Choose stock of a country feature and analyse indicator automatic
feature paused and removed. Will restart after the defense.